Remove `Args` from the `Config` struct
This was convenient, but doesn't make a lot of sense
going forward. Considering that the goal is to expose
the functionality of `rly` as a library in addition to
a CLI, it would be quite cumbersome for a library user
to populate the `Args` struct in order to instantiate
`Config`.